backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
openai==1.3.5
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.2
